create a filter component WIP
the code needs further refactoring. there's an issue with filters dismissing all the motions which must be fixed
